+++ .iex.exs
@@ -1,4 +1,4 @@
 IEx.configure(
-  inspect: [base: :decimal, charlists: :as_lists, binaries: :as_binaries],
+  inspect: [base: :decimal, charlists: :as_lists],
   auto_reload: true
 )